requirements.txt
httpx>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
PyYAML>=6.0
pytest>=7.0.0
langgraph>=0.0.49
